Rework automations UI to match Codex: Current/Paused list + detail view
- Split the Automations page into Current / Paused sections of compact,
  click-through rows (status dot, name, context subtitle, right-aligned cadence
  or "Paused"), matching the Codex app layout.
- Add a dedicated automation detail route (/automations/$automationId) with a
  breadcrumb, Run now / Pause / Edit / Delete actions, the prompt, a Status rail
  (status, next run, last ran), a Details rail (runs-in, project, repeats, mode,
  model, max iterations), and a Previous runs list with open-thread and cancel.
- Extract shared types, formatters (cadence, relative time, status dot), the
  create/edit dialog, and a useAutomations() data+mutations hook into
  -automations.shared.tsx so the list and detail routes share one source of truth.
